Logic Outputs: Rising & falling slopes implemented.

+++ src/gui/circuitwidget/components/logic/adc.cpp
@@ -70,7 +70,7 @@ void ADC::stamp()
 void ADC::voltChanged()
 {
     double volt = m_inPin[0]->getVolt();
-    m_nextOutVal = (int)(volt*m_maxValue/m_maxVolt+0.1);
+    m_nextOutVal = volt*m_maxValue/m_maxVolt+0.1;
     LogicComponent::sheduleOutPuts( this );
 }
